Move configuration up to index to make it more visible, clean up unused feefactory
