+++ src/composition/Types.ts
@@ -1,17 +1,11 @@
 const TYPES = {
-  ArchiveBlockDataProvider: Symbol('ArchiveBlockDataProvider'),
-  ArchiveStakesDataProvider: Symbol('ArchiveStakesDataProvider'),
   BlockDataProviderFactory: Symbol('BlockDataProviderFactory'),
-  MinaExplorerDataProvider: Symbol('MinaExplorerDataProvider'),
-  MinaExplorerStakeDataProvider: Symbol('MinaExplorerStakeDataProvider'),
   StakeDataProviderFactory: Symbol('StakeDataProviderFactory'),
-  FeeCalculatorFactory: Symbol('IFeeCalculatorFactory'),
-  IFeeCalculator: Symbol('IFeeCalculator'),
+  PayoutCalculatorFactory: Symbol('PayoutCalculatorFactory'),
   IAddressRemover: Symbol('IAddressRemover'),
   IBlockProcessor: Symbol('IBlockProcessor'),
   IPaymentBuilder: Symbol('IPaymentBuilder'),
   IPaymentProcessor: Symbol('IPaymentProcessor'),
-  IPayoutCalculator: Symbol('IPayoutCalculator'),
   ITransactionBuilder: Symbol('ITransactionBuilder'),
   ITransactionProcessor: Symbol('ITransactionWriter'),
   ISender: Symbol('ITransactionSender'),
