+++ src/core/transaction/Model.ts
@@ -13,9 +13,3 @@ export interface ITransactionProcessor {
 export interface ISender {
     send(config: PaymentConfiguration, paymentProcess: PaymentProcess): Promise<void>;
 }
-export interface IFeeCalculatorFactory {
-    create(calculatorType: string): IFeeCalculator;
-}
-export interface IFeeCalculator {
-    calculate(transaction: PayoutTransaction, config: PaymentConfiguration): Promise<void>;
-}
